requirements.txt
numpy
pydantic
python-dotenv
google-genai
pinecone
pydantic-settings
colorama
termcolor